various fixes on gdriveStorage and his tests.
removed the removeAttachment method as it was useless.
removed error catching when user was doing request with invalid credentials.
throw error when doing a putAttachment on a folder.

changed varaiables names in test file to respect naming convention.
added/ removed tests following the changes on gdriveStorage file.

added tests verifying that the user can only put "enclosure"
as attachment name in putAttachment and getAttachment.