jiodate: fixes tests in older firefox versions
Tests were failing because some Firefox versions cannot
parse a specific date-string. So the Date constructor
that receives only integers was used. But there is one
details: this contructor uses months from 0 to 11, so if,
for example, we want to have a date in January, we send 0
as month and not 1.