Build: stop relying on grunt, nodejs, npm.
Use a Makefile to concatenate the file (and so, drop grunt).
Allow to build without linting

Fetch external dependencies with curl.
As they are few, it simplifies the process.
Keep a local copy of the dependencies.

Add Jison call in the build process.

JSLint can be also triggered from the makefile, and installed how the developper prefers.

Drop the package.json, as jIO is for now not published on npm website.

Remove unused libraries.