[replicateStorage] Replicate all documents/attachments before raising.
Accumulate all operations done in a ReplicateReport object.

This ReplicateReport is returned by the repair method in any case (success or failure).

If an error occured while replicating a document, do not stop the full process and log the issue in the report.
Which means repair now always tries to replicate all documents even if one fails.

Each operation has a code to ease parsing it if needed.

By default, the report only contains error code (< 100).
It is possible to change the log level by using the 'report_level' storage parameter.

If the 'debug' parameter is set to true, codes will also be displayed in the console.log.